+++ .appveyor.yml
@@ -26,6 +26,6 @@ test_script:
   - pydocstyle
 
   # Test for unit test.
-  - python -m pytest --cov=cube_encryption/ analyzers/ unit_test/
+  - python -m pytest --cov=content/ unit_test/
 
 build: off
