+++ content/analyzers/key_analyzer.py
@@ -3,7 +3,7 @@ import operator
 from functools import reduce
 from typing import List, Optional
 from collections import namedtuple
-from cube_encryption.constants import Key, CUBE_MOVE, COMMUTE_MOVE
+from content.encryption.constants import Key, CUBE_MOVE, COMMUTE_MOVE
 
 
 class KeyAnalyzer:
