+++ content/encryption/cube_face_for_cubie.py
@@ -5,8 +5,8 @@ import numpy as np
 import pandas as pd
 from typing import List
 from collections import deque
-from cube_encryption.cubie import Cubie
-from cube_encryption.constants import CUBIE_LENGTH, WRONG_SIDE_LENGTH, \
+from content.encryption.cubie import Cubie
+from content.encryption.constants import CUBIE_LENGTH, WRONG_SIDE_LENGTH, \
     WRONG_CUBE_FACE_INPUT, WRONG_FRAME_INDEX_NAME, WRONG_FRAME_COLUMN_NAME, \
     CubieItem
