+++ content/encryption/cube_for_cubie.py
@@ -2,8 +2,8 @@
 
 import numpy as np
 from typing import List
-from cube_encryption.cube_face_for_cubie import CubeFaceForCubie
-from cube_encryption.constants import WRONG_CUBE_INPUT, CUBIE_LENGTH, \
+from content.encryption.cube_face_for_cubie import CubeFaceForCubie
+from content.encryption.constants import WRONG_CUBE_INPUT, CUBIE_LENGTH, \
     CubeMove, WRONG_CUBE_MOVE, WRONG_CUBE_SIDE_LENGTH, CubieItem, Key
